Replce calculator's input and output from double to int for exception test

+++ spring-intro/src/main/java/com/xiaoming/spring/aop_basic/dynamic_proxy_intro/CalculatorImpl.java
@@ -3,23 +3,23 @@ package com.xiaoming.spring.aop_basic.dynamic_proxy_intro;
 import com.xiaoming.spring.aop_basic.traditional_impl.Calculator;
 
 public class CalculatorImpl implements Calculator {
-    public double add(double a, double b) {
-        double result = a + b;
+    public int add(int a, int b) {
+        int result = a + b;
         return result;
     }
 
-    public double sub(double a, double b) {
-        double result = a - b;
+    public int sub(int a, int b) {
+        int result = a - b;
         return result;
     }
 
-    public double mul(double a, double b) {
-        double result = a * b;
+    public int mul(int a, int b) {
+        int result = a * b;
         return result;
     }
 
-    public double div(double a, double b) {
-        double result = a / b;
+    public int div(int a, int b) {
+        int result = a / b;
         return result;
     }
 }
